[#3736] Should be fixed now

+++ nes-runtime/src/Execution/Operators/Streaming/Join/StreamHashJoin/DataStructure/StreamHashJoinWindow.cpp
@@ -31,7 +31,7 @@ Operators::StreamJoinHashTable* StreamHashJoinWindow::getHashTable(bool leftSide
     }
 }
 
-size_t StreamHashJoinWindow::getNumberOfTuplesOfWorker(bool isLeftSide, uint64_t workerIdx) {
+uint64_t StreamHashJoinWindow::getNumberOfTuplesOfWorker(bool isLeftSide, uint64_t workerIdx) {
     return getHashTable(isLeftSide, workerIdx)->getNumberOfTuples();
 }
